requirements.txt
flask
firebase-admin
openai
requests
pdfplumber
scipy
python-dotenv